Created a filter by limiting feature
Implemented the the ability to filter by limiting the number of todos shown by adding it as a parameter in the getTodos method in the todo.service. Also wrote the complimentary code in the getTodos method as well as the todo.component.html to make it run smoothly on the local host. Filtered on the server vs the client.